fix(ci): split files over quality limits
Amp-Thread-ID: https://ampcode.com/threads/T-01a02b27-8731-733e-b188-5e481a47eee8
